requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
PyYAML==6.0
soupsieve==2.3.1
